# overload mul for PermMatrix