not necessary to overload * twice

Before

```julia
julia> @benchmark $H * $X
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   90.942 μs …  1.097 ms  ┊ GC (min … max): 0.00% … 78.45%
 Time  (median):      97.175 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   104.297 μs ± 54.600 μs  ┊ GC (mean ± σ):  3.66% ±  6.46%

  ▂▇█▆▅▄▂▁▁▁                                                   ▂
  ███████████▆▆▆▅▃▁▄▁▄▃▁▁▃▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▃▃▅▇███ █
  90.9 μs       Histogram: log(frequency) by time       233 μs <

 Memory estimate: 256.05 KiB, allocs estimate: 2.
```

After

```
julia> @benchmark $H * $X
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   87.427 μs …  1.272 ms  ┊ GC (min … max): 0.00% … 81.04%
 Time  (median):     100.521 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   108.521 μs ± 67.480 μs  ┊ GC (mean ± σ):  4.46% ±  6.63%

    ▃▆██▇▇▃▂▁▁                                              ▁▁ ▂
  ▃█████████████▇▇▇▆▆▄▆▇▄▄▁▅▄▄▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▇██ █
  87.4 μs       Histogram: log(frequency) by time       233 μs <

 Memory estimate: 256.05 KiB, allocs estimate: 2.

```